Include auto generated changes made by server
When MC server is run the config files are modified to inlcude comments,
to have a particular whitespacing and to have options in alphabetical order.

In order to allow easier diffs in the future from running configuration files
this change adopts the changes that the server is making automatically.

There should be no functional changes included just orderingin and formating
changes